added grid system, add scene manager and lost scene

+++ Assets/Snake/snake.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "Node.h"
 #include <unordered_map>
+
 enum directions
 {
 	left = 1,
@@ -12,11 +13,11 @@ class snake
 {
 private:
 	int counter = 0;
-	std::unordered_map<directions, sf::Vector2f> directionVecs;
+	std::unordered_map<directions, sf::Vector2i> directionVecs;
 	Node* SnakeHead;
 	Node* SnakeTail;
 	float speed;
-	sf::Vector2f currentDirection;
+	sf::Vector2i currentDirection;
 	float GetDistance(const sf::Vector2f& pos1, const sf::Vector2f& pos2);
 public:
 	snake();
